reselect использован для фильтрации
