+++ src/store/reducers/user-interface/selectors.js
@@ -1,3 +1,4 @@
 export const getActiveCityName = (state) => state.INTERFACE.city.name;
 export const getActiveCityCoords = (state) => state.INTERFACE.city.coordinates;
 export const getHoveredCard = (state) => state.INTERFACE.hoveredCard;
+export const getActiveSorting = (state) => state.INTERFACE.activeSorting;
